Fix panic if file not found #1

+++ src/main.rs
@@ -1,6 +1,6 @@
 use clap::Parser;
-use std::fs;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
+use std::{fs, process};
 
 #[derive(Parser)]
 #[command(version, about, long_about = None)]
@@ -13,14 +13,31 @@ struct Opts {
     filename: PathBuf,
 }
 
+fn check_file_exists(file_path: &str) -> bool {
+    match Path::new(file_path).metadata() {
+        Ok(_) => true,
+        Err(_) => false,
+    }
+}
+
 fn backup_file(file_name: &str) {
-    let bak_name = format!("{}.bak", file_name);
-    fs::copy(file_name, &bak_name).expect("Failed to create backup file");
+    if check_file_exists(file_name) {
+        let bak_name = format!("{}.bak", file_name);
+        fs::copy(file_name, &bak_name).expect("Failed to create backup file");
+    } else {
+        eprintln!("File {} not found", file_name);
+        process::exit(1);
+    }
 }
 
 fn delete_backup_file(file_name: &str) {
     let bak_name = format!("{}.bak", file_name);
-    fs::remove_file(&bak_name).expect("Failed to delete backup file");
+    if check_file_exists(&bak_name) {
+        fs::remove_file(&bak_name).expect("Failed to delete backup file");
+    } else {
+        eprintln!("File {} not found", bak_name);
+        process::exit(1);
+    }
 }
 
 fn main() {
